+++ .github/pull_request_template.md
@@ -30,6 +30,6 @@ if __name__ == '__main__':
 ## Logseq
 
 - [ ] Make logseq notes and flash cards.
-- [ ] Do not use logseq aliases so that the graph looks clean and its more navigable.
 - [ ] Use singular nouns for tags.
-- [ ] Use spaces in filenames instead of `-` or `_` just so that you don't have to use aliases (ugly I know).
+- [ ] Use `-` or `_` in filenames instead of spaces.
+- [ ] Use aliases for spaces and plurals.
