+++ notes/pages/duelingdqn.md
@@ -1,0 +1,3 @@
+---
+alias: dueling deep q network
+---
