+++ notes/pages/log-probability.md
@@ -1,0 +1,3 @@
+---
+alias: log probability, log prob
+---
