+++ notes/pages/policy-gradient.md
@@ -1,0 +1,3 @@
+---
+alias: policy gradient, pg
+---
