--- notes/pages/value based methods.md
@@ -1,1 +1,0 @@
--
