+++ notes/pages/value-based-methods.md
@@ -1,0 +1,3 @@
+---
+alias: value-based methods, value based methods, value methods
+---
